requirements.txt
streamlit==1.32.0
google-cloud-texttospeech==2.16.1
google-auth==2.28.1
pydub==0.25.1
python-dotenv==1.0.1
